Make sure plugin API is available.
The videojs_rails gem used to use the dev version of video JS in 4.6.1,
4.6.4 changed to use the minified version.

We’re having problems running the minified version on FutureLearn.com 
against our plugins, particularly one that adds a button to the
control bar.

It seems that it’s an ongoing problem in the video js project that
the minified version doesn’t correctly expose all the expected APIs
that plugins need and use[1][2].

We haven’t been able to confirm what API method is unavailable- it
looks to be something around overriding behaviour of elements. There
were no errors in the console so it was hard to pin down what is wrong.
We’ve tried other plugins that add buttons that are listed on the video
js project (such as the loop button[3]) and have had the same problem.

We feel that using the dev version is the easiest way to work around
this problem until video js improves their minification process[1] and
we don’t lose much since most rails projects will be using the asset
pipeline to perform minification.


[1] https://github.com/videojs/video.js/issues/1586
[2] https://github.com/videojs/video.js/issues/1491
[3] https://github.com/CharlotteDunois/videojs-loopbutton