Don’t force users of gem to rely on a CDN for serving video player SWF.
We don’t want to depend on a CDN that may be down or be blocked
in other countries. The more CDNs and services depended on the more
likely that one of them may have problems.

This commit introduces the SWF back in to the gem that was removed
in [1]. Users of the gem will now be able to override video js’s
default behaviour in their own project by following the instructions
in [2].

We could work around our need to not use the CDN version of the SWF
file by including it in our project, but since the version of the SWF 
should be kept in sync with the version of video js I think it makes
more sense for it to live in the gem.

[1] https://github.com/seanbehan/videojs_rails/commit/65d7aa707984f2798d7d172d1b8b90df41bc0c98
[2] https://github.com/videojs/video.js/blob/master/docs/guides/setup.md#self-hosted