Updated readme.md
Added info about an alternative way to set the Flash SWF file location. I couldn't find any place in video.js to set the flash player SWF file location, so it seems the approach I added is currently the only one. If this is true, you might want to remove the suggestion of setting the location in the video.js file directly.